+++ events.js
@@ -162,7 +162,52 @@ export const events = {
         effect: function(){
             return tax_revolt();
         }
-    }
+    },
+    slave_death1: {
+        reqs: { 
+            trait: 'slaver',
+            tech: 'slaves'
+        },
+        effect: function(){
+            if (global.city['slave_pen'] && global.city.slave_pen.slaves > 0){
+                global.city.slave_pen.slaves--;
+                return loc('event_slave_death1');
+            }
+            else {
+                return loc('event_slave_none');
+            }
+        }
+    },
+    slave_death2: {
+        reqs: { 
+            trait: 'slaver',
+            tech: 'slaves'
+        },
+        effect: function(){
+            if (global.city['slave_pen'] && global.city.slave_pen.slaves > 0){
+                global.city.slave_pen.slaves--;
+                return loc('event_slave_death2');
+            }
+            else {
+                return loc('event_slave_none');
+            }
+        }
+    },
+    slave_death3: {
+        reqs: { 
+            trait: 'slaver',
+            tech: 'slaves'
+        },
+        effect: function(){
+            if (global.city['slave_pen'] && global.city.slave_pen.slaves > 0){
+                global.city.slave_pen.slaves--;
+                return loc('event_slave_death3');
+            }
+            else {
+                return loc('event_slave_none');
+            }
+        }
+    },
 };
 
 function tax_revolt(){